+++ packages/fighting-design/alert/src/alert.vue
@@ -2,6 +2,7 @@
   import { Props, Emits } from './alert'
   import { computed, ref } from 'vue'
   import { FIcon } from '../../icon'
+  import { sizeChange } from '../../_utils'
   import type { ComputedRef, CSSProperties, Ref } from 'vue'
   import type { handleCloseInterface as a } from './interface'
   import type { classListInterface as b } from '../../_interface'
@@ -39,18 +40,15 @@
   })
 
   const styleList: ComputedRef<CSSProperties> = computed((): CSSProperties => {
-    const { fontSize, color, background, fixed, fixedStyle } = prop
+    const { fontSize, color, background, titleSize, titleColor } = prop
 
-    const style = {
-      color,
-      background,
-      fontSize
-    } as const
-
-    if (fixed && fixedStyle) {
-      return { ...style, ...fixedStyle } as const
+    return {
+      '--f-alert-color': color,
+      '--f-alert-title-color': titleColor,
+      '--f-alert-background': background,
+      '--f-alert-font-size': sizeChange(fontSize),
+      '--f-alert-title-size': sizeChange(titleSize)
     }
-    return style
   })
 
   const handleClose: a = (evt: MouseEvent): void => {
@@ -65,9 +63,11 @@
       <f-icon :icon="icon" />
       {{ title }}
     </div>
+
     <div v-if="$slots.default" :class="subTitleClassList">
       <slot />
     </div>
+
     <span v-if="close" class="f-alert-close" @click.stop="handleClose">
       <f-icon icon="f-icon-close" />
     </span>