+++ packages/fighting-design/alert/src/interface.d.ts
@@ -5,10 +5,3 @@ export type overflowType = 'hidden' | 'roll' | 'ellipsis' | ''
 export interface handleCloseInterface {
   (evt: MouseEvent): void
 }
-
-export interface fixedStyleInterface {
-  top?: string
-  button?: string
-  right?: string
-  left?: string
-}