+++ start/src/App.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup>
-  import { mdPlugin } from './utils/index'
-  import { md } from './utils/md'
+  // import { mdPlugin } from './utils/index'
+  // import { md } from './utils/md'
 
   // console.log(md)
-  const mdRes = mdPlugin.render(md)
+  // const mdRes = mdPlugin.render(md)
   // console.log(mdRes)
 
   // const res = mdPlugin(md)
@@ -11,7 +11,9 @@
 </script>
 
 <template>
-  <f-image width="200px" height="400px" :round="50" caption="哈哈哈哈哈哈哈" src="https://tianyuhao.cn/images/auto/1.jpg" />
+  <f-alert title="你好" type="primary" title-color="red" close>
+    默认提示信息
+  </f-alert>
 
   <!-- <div v-html="mdRes"></div> -->
 </template>